Simplify configuration of workers
